refactor: rename `games` tab to `library`
Also rename the following `library` subfolders
* `game_info` -> `details`
* `games_widgets` -> `widgets`
