+++ rare/components/tabs/library/details/__init__.py
