+++ rare/components/tabs/library/details/details.py
