+++ rare/components/tabs/library/details/dlcs.py
