+++ rare/components/tabs/library/details/settings.py
