+++ rare/components/tabs/library/widgets/icon_game_widget.py
