+++ rare/components/tabs/library/widgets/library_widget.py
