+++ rare/resources/static_css/__init__.py
@@ -70,7 +70,7 @@ qt_resource_struct = b"\
 \x00\x00\x00\x00\x00\x02\x00\x00\x00\x01\x00\x00\x00\x02\
 \x00\x00\x00\x00\x00\x00\x00\x00\
 \x00\x00\x00\x1a\x00\x01\x00\x00\x00\x01\x00\x00\x00\x00\
-\x00\x00\x01\x906&\x0a\x17\
+\x00\x00\x01\x92\x09g[\xe6\
 "
 
 def qInitResources():
